GameScreen now supports scrolling. Added "hover-listening", but undecided about the best way to maintain that.
GameScreen provides a tree of two Actors: a SparseLayers (being the actual game-map display component) and a Container.
The Container fills the window and clips the SparseLayers to fit.
The SparseLayers is positioned so the desired "camera-point" or "scroll-point" is at the Container's (0,0).

We have the ability to specify hover-zones on GameScreen. This specific functionality may be abstracted into a helper-class or something.